data-engineering/data-pipe-line/requirements.txt
PyMySQL==1.0.2
python-dotenv==0.20.0
redis==4.3.1
schedule==1.1.0
streamlit==1.9.0
voila==0.3.5
wandb==0.12.14
PyYAML==6.0
pymongo==4.1.1
pymongo[srv]
